requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.2
django-appconf==1.0.3
django-avatar==4.1.0
django-crispy-forms==1.7.2
django-embed-video==1.2.0
django-filter==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.10.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
markdown2==2.3.8
mccabe==0.6.1
Pillow==6.2.0
psycopg2==2.8.3
pylint==2.3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==5.1.0
wrapt==1.11.2
